# Fix performance issue in RingUtils::checkFused